create class details panel for admins
