+++ frontend/src/components/ClassScheduleCard/index.css
@@ -1,0 +1,20 @@
+.class-schedule-card {
+    padding-bottom: 20px;
+    border-bottom: 1px solid #e3e3e3;
+}
+
+.schedule-header {
+    position: relative;
+    padding-bottom: 20px;
+    display: flex;
+    flex-direction: column;
+    gap: 4px;
+}
+
+.schedule-header-day {
+    font-family: 'Montserrat';
+    font-weight: 600;
+    font-size: 18px;
+    color: #0f1111;
+    margin-bottom: 6px;
+}
